unsharing the handler core logic
